// src/app/model/user-profile.ts
export interface UserProfile {
  firstname: string;
  lastname: string;
  email: string;
  mobile: string;
  dob: string;
  updatedDate: string;
  createdDate: string;
  address: string;
  profilepic: string;
  location: { city: string; state: string };
  userid: string;
  accountType: string;
  id: string;
}